requirements.txt
# Web Framework & Server
flask
gunicorn

# AWS SDK (Must be 1.28.57+ for Bedrock support)
boto3>=1.28.57
python-dotenv

# Search & Database
opensearch-py
requests-aws4auth
requests

# Document Processing
pypdf